LU-5577 obd: change brw_page->count to unsigned
Pages count is unsigned. So, change the type accordant usage.

Reviewed-on: http://review.whamcloud.com/12378
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/osc/osc_cl_internal.h
@@ -74,7 +74,7 @@ struct osc_io {
 	 * page writeback from happening. */
 	struct osc_extent *oi_trunc;
 
-	int oi_lru_reserved;
+	unsigned long oi_lru_reserved;
 
 	struct obd_info    oi_info;
 	struct obdo        oi_oa;
